chore(lld): refactor LNSBanner related types
