+++ .changeset/wise-swans-rule.md
@@ -1,0 +1,5 @@
+---
+"ledger-live-desktop": minor
+---
+
+Improve LNSBanner tests and code quality
