+++ apps/ledger-live-desktop/src/newArch/features/LNSUpsell/components/LNSUpsellBanner/types.ts
@@ -1,0 +1,13 @@
+import { Icons } from "@ledgerhq/react-ui";
+
+export type LNSBannerModel = {
+  variant: LNSBannerVariant;
+  discount?: number;
+  tracking: "opted_in" | "opted_out";
+  handleCTAClick: () => void;
+};
+
+export type LNSBannerVariant =
+  | { type: "none" }
+  | { type: "banner"; image?: string }
+  | { type: "notification"; icon: keyof typeof Icons };
