--- apps/ledger-live-desktop/src/newArch/features/LNSUpsell/types/enum/Analytics.ts
@@ -1,11 +1,0 @@
-export enum AnalyticsPage {
-  Accounts = "Accounts",
-  Manager = "Manager",
-  NotificationPanel = "Notification Panel",
-  Portfolio = "Portfolio",
-}
-
-export enum AnalyticsButton {
-  CTA = "Level up wallet",
-  LearnMore = "learn more",
-}
