+++ apps/ledger-live-desktop/src/newArch/features/LNSUpsell/types/index.ts
@@ -1,14 +1,4 @@
-import { Icons } from "@ledgerhq/react-ui";
 import { Feature_LldNanoSUpsellBanners } from "@ledgerhq/types-live";
-
-export * from "./enum/Analytics";
-
-export type LNSBannerModel = {
-  variant: LNSBannerVariant;
-  discount?: number;
-  tracking: Tracking;
-  handleCTAClick: () => void;
-};
 
 export type LNSBannerState = {
   isShown: boolean;
@@ -20,11 +10,6 @@ export type LNSBannerLocation = Extract<
   "manager" | "accounts" | "notification_center" | "portfolio",
   keyof FFParams["opted_in"] | keyof FFParams["opted_out"]
 >;
-
-export type LNSBannerVariant =
-  | { type: "none" }
-  | { type: "banner"; image?: string }
-  | { type: "notification"; icon: keyof typeof Icons };
 
 type Tracking = "opted_in" | "opted_out";
 type FFParams = Required<Feature_LldNanoSUpsellBanners>["params"];
